Refactor: Split xtce-generator output
Split generator output into two seperate files for telemetry and commanding
